third_party/WebKit/LayoutTests/platform/win/virtual/scalefactor200withzoom/fast/hidpi/static/gesture-scroll-amount-expected.txt

This test ensures that touch gesture scrolls the right amount pixels in high dpi mode.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS successfullyParsed is true

TEST COMPLETE
first gesture
PASS document.scrollingElement.scrollTop is 1
PASS document.scrollingElement.scrollLeft is 1
second gesture
PASS document.scrollingElement.scrollTop is 3
PASS document.scrollingElement.scrollLeft is 3
PASS successfullyParsed is true

TEST COMPLETE
